requirements.txt
streamlit
langchain
langchain-groq
langgraph
requests
tavily-python
python-dotenv
langdetect
Pillow
pytesseract
pdf2image
pypdf
faster-whisper
numpy
pydantic
PyPDF2
bs4
langchain_community
sentence-transformers
chromadb
pysqlite3-binary
